+++ src/render/skybox.rs
@@ -1,5 +1,4 @@
-use crate::io::Color;
-use crate::utils::Ray;
+use crate::utils::{Color, Ray};
 
 pub trait SkyBox {
     fn get_color(&self, r: Ray) -> Color;
@@ -13,7 +12,7 @@ pub struct ColorGradientSkyBox {
 impl SkyBox for ColorGradientSkyBox {
     fn get_color(&self, r: Ray) -> Color {
         let unit = r.direction().unit_vector();
-        let t = 0.5 * (unit.y() + 1.0);
+        let t = (0.5 * (unit.y() + 1.0)) as f32;
         self.v1 * (1.0 - t) + self.v2 * t
     }
 }
